requirements.txt
discord.py[voice]
pymongo
python-dotenv
motor[srv]